Check whether m_toplevel exists before dereferencing it
The parent surface could be a popup. In that case m_toplevel would be
nullptr. Since xdgshell doesn't allow an xdg_popup to be the parent of
an xdg_toplevel, the best we can do is pretend there is no parent in
this case.

Fixes: QTBUG-130128
Change-Id: Ib0367429b71ee96cbd1c3119e77769c511370624
Reviewed-by: Vlad Zahorodnii <vlad.zahorodnii@kde.org>
(cherry picked from commit 9c00d532b1d7fbca8da658bc8a3bb2c53a47ae29)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>